+++ packages/ua-utils-evm-hardhat-test/test/task/getDefaultConfig.test.ts
@@ -1,6 +1,7 @@
 import { defaultExecutorConfig, defaultUlnConfig, setupDefaultEndpoint } from '../__utils__/endpoint'
 import { createContractFactory, getEidForNetworkName } from '@layerzerolabs/utils-evm-hardhat'
 import hre from 'hardhat'
+import { TASK_LZ_GET_DEFAULT_CONFIG } from '@layerzerolabs/ua-utils-evm-hardhat'
 
 describe('task: getDefaultConfig', () => {
     beforeEach(async () => {
@@ -9,7 +10,7 @@ describe('task: getDefaultConfig', () => {
 
     it('should return default configurations', async () => {
         const networks = Object.keys(hre.userConfig.networks ?? {})
-        const getDefaultConfigTask = await hre.run('getDefaultConfig', { networks: networks.toString() })
+        const getDefaultConfigTask = await hre.run(TASK_LZ_GET_DEFAULT_CONFIG, { networks: networks.toString() })
         const contractFactory = createContractFactory()
         for (const localNetwork of networks) {
             const localEid = getEidForNetworkName(localNetwork)
