+++ packages/ua-utils-evm-hardhat-test/test/task/getOAppConfig.test.ts
@@ -2,6 +2,7 @@ import { defaultExecutorConfig, defaultUlnConfig, setupDefaultEndpoint } from '.
 import { createContractFactory, getEidForNetworkName } from '@layerzerolabs/utils-evm-hardhat'
 import hre from 'hardhat'
 import { AddressZero } from '@ethersproject/constants'
+import { TASK_LZ_GET_OAPP_CONFIG } from '@layerzerolabs/ua-utils-evm-hardhat'
 
 describe('task: getOAppConfig', () => {
     beforeEach(async () => {
@@ -11,7 +12,7 @@ describe('task: getOAppConfig', () => {
     it('should return app default configurations when addresses are not oapps', async () => {
         const networks = Object.keys(hre.userConfig.networks ?? {})
         const addresses = new Array(networks.length).fill(AddressZero).toString()
-        const getDefaultConfigTask = await hre.run('getOAppConfig', {
+        const getDefaultConfigTask = await hre.run(TASK_LZ_GET_OAPP_CONFIG, {
             networks: networks.toString(),
             addresses: addresses.toString(),
         })
