+++ packages/ua-utils-evm-hardhat/src/constants/tasks.ts
@@ -1,1 +1,3 @@
 export const TASK_LZ_WIRE_OAPP = 'lz:oapp:wire'
+export const TASK_LZ_GET_DEFAULT_CONFIG = 'lz:oapp:getDefaultConfig'
+export const TASK_LZ_GET_OAPP_CONFIG = 'lz:oapp:getOAppConfig'
