+++ packages/ua-utils-evm-hardhat/src/tasks/index.ts
@@ -1,3 +1,1 @@
 import './oapp'
-import './getDefaultConfig'
-import './getOAppConfig'
