+++ packages/ua-utils-evm-hardhat/src/tasks/oapp/getDefaultConfig.ts
@@ -1,33 +1,24 @@
 import { ActionType } from 'hardhat/types'
 import { task } from 'hardhat/config'
 import { getReceiveConfig, getSendConfig, printConsoleTable } from '@/utils/taskHelpers'
-import assert from 'assert'
+import { TASK_LZ_GET_DEFAULT_CONFIG } from '@/constants'
 
 interface TaskArgs {
     networks: string
-    addresses: string
 }
 
-export const getOAppConfig: ActionType<TaskArgs> = async (taskArgs, ...args) => {
-    const networks = taskArgs.networks.split(',')
-    const addresses = taskArgs.addresses.split(',')
-    assert(networks.length === addresses.length, 'Passed in networks must match length of passed in addresses.')
+export const getDefaultConfig: ActionType<TaskArgs> = async (taskArgs) => {
+    const networks = new Set(taskArgs.networks.split(','))
     const configs: Record<string, Record<string, unknown>> = {}
-
-    for (const [index, localNetworkName] of networks.entries()) {
+    for (const localNetworkName of networks) {
         configs[localNetworkName] = {}
         for (const remoteNetworkName of networks) {
             if (remoteNetworkName === localNetworkName) continue
             const [sendLibrary, sendUlnConfig, sendExecutorConfig] = await getSendConfig(
                 localNetworkName,
-                remoteNetworkName,
-                addresses[index]
+                remoteNetworkName
             )
-            const [receiveLibrary, receiveUlnConfig] = await getReceiveConfig(
-                localNetworkName,
-                remoteNetworkName,
-                addresses[index]
-            )
+            const [receiveLibrary, receiveUlnConfig] = await getReceiveConfig(localNetworkName, remoteNetworkName)
 
             configs[localNetworkName][remoteNetworkName] = {
                 defaultSendLibrary: sendLibrary,
@@ -52,9 +43,8 @@ export const getOAppConfig: ActionType<TaskArgs> = async (taskArgs, ...args) => 
 }
 
 task(
-    'getOAppConfig',
+    TASK_LZ_GET_DEFAULT_CONFIG,
     'outputs the default Send and Receive Messaging Library versions and the default application config'
 )
     .addParam('networks', 'comma separated list of networks')
-    .addParam('addresses', 'comma separated list of addresses')
-    .setAction(getOAppConfig)
+    .setAction(getDefaultConfig)
