+++ packages/ua-utils-evm-hardhat/src/tasks/oapp/getOAppConfig.ts
@@ -1,23 +1,34 @@
 import { ActionType } from 'hardhat/types'
 import { task } from 'hardhat/config'
 import { getReceiveConfig, getSendConfig, printConsoleTable } from '@/utils/taskHelpers'
+import { TASK_LZ_GET_OAPP_CONFIG } from '@/constants/tasks'
+import assert from 'assert'
 
 interface TaskArgs {
     networks: string
+    addresses: string
 }
 
-export const getDefaultConfig: ActionType<TaskArgs> = async (taskArgs) => {
-    const networks = new Set(taskArgs.networks.split(','))
+export const getOAppConfig: ActionType<TaskArgs> = async (taskArgs, ...args) => {
+    const networks = taskArgs.networks.split(',')
+    const addresses = taskArgs.addresses.split(',')
+    assert(networks.length === addresses.length, 'Passed in networks must match length of passed in addresses.')
     const configs: Record<string, Record<string, unknown>> = {}
-    for (const localNetworkName of networks) {
+
+    for (const [index, localNetworkName] of networks.entries()) {
         configs[localNetworkName] = {}
         for (const remoteNetworkName of networks) {
             if (remoteNetworkName === localNetworkName) continue
             const [sendLibrary, sendUlnConfig, sendExecutorConfig] = await getSendConfig(
                 localNetworkName,
-                remoteNetworkName
+                remoteNetworkName,
+                addresses[index]
             )
-            const [receiveLibrary, receiveUlnConfig] = await getReceiveConfig(localNetworkName, remoteNetworkName)
+            const [receiveLibrary, receiveUlnConfig] = await getReceiveConfig(
+                localNetworkName,
+                remoteNetworkName,
+                addresses[index]
+            )
 
             configs[localNetworkName][remoteNetworkName] = {
                 defaultSendLibrary: sendLibrary,
@@ -42,8 +53,9 @@ export const getDefaultConfig: ActionType<TaskArgs> = async (taskArgs) => {
 }
 
 task(
-    'getDefaultConfig',
+    TASK_LZ_GET_OAPP_CONFIG,
     'outputs the default Send and Receive Messaging Library versions and the default application config'
 )
     .addParam('networks', 'comma separated list of networks')
-    .setAction(getDefaultConfig)
+    .addParam('addresses', 'comma separated list of addresses')
+    .setAction(getOAppConfig)
