+++ packages/ua-utils-evm-hardhat/src/tasks/oapp/index.ts
@@ -1,1 +1,3 @@
 import './wire'
+import './getDefaultConfig'
+import './getOAppConfig'
